HBX-1990: Move class 'IndexProcessor' from 'org.hibernate.tool.internal.reveng' to 'org.hibernate.tool.internal.reveng.reader'
